% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_do_computations.R
\name{rf_do_computations}
\alias{rf_do_computations}
\title{Perform computations on data}
\usage{
rf_do_computations(computation_script, records_data)
}
\arguments{
\item{computation_script}{Character. Path to computation script}

\item{records_data}{Data frame of data}
}
\value{
Computed data object.
}
\description{
Runs user-defined computations on the raw data prior to rendering.
}
