% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_generate.R
\name{rf_generate}
\alias{rf_generate}
\title{Render feedback content for all recipients}
\usage{
rf_generate(batch_id)
}
\arguments{
\item{batch_id}{Character. Identifier for the batch.}
}
\value{
Invisible
}
\description{
Generates personalized feedback documents for each recipient.
}
