% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_get_recipients.R
\name{rf_get_recipients}
\alias{rf_get_recipients}
\title{Load recipient data}
\usage{
rf_get_recipients()
}
\value{
Invisible
}
\description{
Loads or updates the recipients file as specified in the config.
}
