% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_init.R
\name{rf_init}
\alias{rf_init}
\title{Initialise a recorder feedback project}
\usage{
rf_init(path = ".")
}
\arguments{
\item{path}{Character. Path to create the project.}
}
\value{
Invisible
}
\description{
Sets up the directory structure and template files for a new feedback project.
}
