% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_look_for_computed_objects.R
\name{rf_look_for_computed_objects}
\alias{rf_look_for_computed_objects}
\title{Look in the R Markdown template and identify computed objects required in the code to render the template.}
\usage{
rf_look_for_computed_objects(file_path_to_template)
}
\arguments{
\item{file_path_to_template}{The file path to the R Markdown template.}
}
\value{
A list containing two elements:
\itemize{
  \item \code{bg_computed_objects}: A character vector representing unique computed objects following "params$bg_computed_objects$".
  \item \code{user_computed_objects}: A character vector representing unique computed objects following "params$user_computed_objects$".
}
}
\description{
This function reads an R Markdown template file and identifies computed objects used in the code. It specifically looks for background computed objects (`params$bg_computed_objects$`) and user computed objects (`params$user_computed_objects$`).
}
