% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_render_content.R
\name{rf_render_content}
\alias{rf_render_content}
\title{Render feedback content for all recipients}
\usage{
rf_render_content(
  template_file,
  recipient_params,
  recipient_id,
  batch_id,
  email_format,
  template_html
)
}
\arguments{
\item{template_file}{Character. RMarkdown file for template}

\item{recipient_params}{List. Parameters for the recipient}

\item{recipient_id}{Character. Identifier for the batch.}

\item{batch_id}{Character. Identifier for the batch.}

\item{template_html}{Character. Template file for template}
}
\value{
Path to rendered file.
}
\description{
Generates personalized feedback documents for each recipient.
}
