% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_render_single.R
\name{rf_render_single}
\alias{rf_render_single}
\title{Render feedback for a single recipient}
\usage{
rf_render_single(recipient_id, view = F)
}
\arguments{
\item{recipient_id}{Integer or character. ID of the recipient.}

\item{view}{Logical. Whether to show the content in the viewer.}
}
\value{
Path to rendered file.
}
\description{
Generates feedback for one recipient using their data.
}
