% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_dispatch_smtp.R
\name{rf_dispatch_smtp}
\alias{rf_dispatch_smtp}
\title{Dispatch feedback via SMTP email}
\usage{
rf_dispatch_smtp(batch_id)
}
\arguments{
\item{batch_id}{Character. Identifier for the batch.}
}
\value{
Message indicating status.
}
\description{
Sends generated feedback files to recipients using SMTP.
}
